refactor(app): extract LlmSummaryCard derivation into composable (#397)
Move the SummaryJson/DerivedConfidence types and the setup() computed logic out
of LlmSummaryCard.vue (706 -> 392 lines) into useLlmSummaryCard.ts. The card
stays a thin presentation shell over admin-generated cached LLM summaries; all
AI-disclosure copy, verdict labels, and tooltips are unchanged.

Advances #397. Part of #346.
